requirements.txt
#python=3.13
pygame-ce==2.5.5
pygbag==0.9.2